pyproject.toml
[tool.poetry]
name = "grpc-text-app"
version = "0.1.0"
description = ""
authors = ["Ro"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
grpcio = "^1.70.0"
grpcio-tools = "^1.70.0"
click = "^8.1.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
